requirements.txt
streamlit
watchdog

huggingface_hub
pillow
python-dotenv
